complete hero
