+++ src/components/layout/footer.js
@@ -369,10 +369,10 @@ const Footer = () => (
                             <Title>{localize('PARTNER WITH US')}</Title>
                         </div>
                         <div>
-                            <Link to="/partners">{localize('Affiliate and IB')}</Link>
+                            <Link to="/partners/affiliate-ib/">{localize('Affiliate and IB')}</Link>
                         </div>
                         <div>
-                            <Link to="/partners/payment-agent">{localize('Payment agent')}</Link>
+                            <Link to="/partners/payment-agent/">{localize('Payment agent')}</Link>
                         </div>
                     </Col>
                     <Col margin_top width="25%">
@@ -462,10 +462,10 @@ const Footer = () => (
                         header_style={mobile_accordion_header}
                     >
                         <Item>
-                            <Link to="/partners">{localize('Affiliate and IB')}</Link>
+                            <Link to="/partners/affiliate-ib/">{localize('Affiliate and IB')}</Link>
                         </Item>
                         <Item>
-                            <Link to="/partners/payment-agent">{localize('Payment agent')}</Link>
+                            <Link to="/partners/payment-agent/">{localize('Payment agent')}</Link>
                         </Item>
                     </AccordionItem>
                 </Accordion>
