+++ src/pages/partners/affiliate-ib/_table.js
