+++ src/pages/partners/payment-agent/index.js
@@ -12,7 +12,7 @@ const PaymentAgent = () => {
     return (
         <Layout type="partners" padding_top="10rem" no_login_signup>
             <SEO
-                title={localize('Partners')}
+                title={localize('Payment agents')}
                 description={localize('Partner with a trusted online trading provider')}
             />
             <Hero />
